fix: resolve cargo fmt and clippy errors breaking CI
All CI runs were failing at `cargo fmt --check` (3 formatting diffs),
with 35 additional clippy warnings-as-errors hidden behind the gate.

+++ krillnotes-core/src/core/workspace/attachments.rs
@@ -431,7 +431,7 @@ impl Workspace {
                 meta.note_id,
                 meta.filename,
                 meta.mime_type,
-                meta.size_bytes as i64,
+                meta.size_bytes,
                 meta.hash_sha256,
                 salt_bytes.as_slice(),
                 meta.created_at,
